api/requirements.txt
requests==2.31.0
pandas==2.2.0
pyarrow==15.0.0
fastparquet==2024.2.0

psycopg2-binary==2.9.9
sqlalchemy==2.0.25
uvicorn==0.22.0
fastapi==0.95.2
python-dotenv==1.0.0
inflection